refactor agenda module into a sub package. fix a bug where sensor scan would not trigger event and a bug where intel was not setting up author properly
